requirements.txt
clashroyale
python-dotenv
aiohttp
motor
dnspython
